test: merge RegenSoloTest and RegenTest
